+++ hs-bindgen/test-artefacts/fixtures/edge-cases/iterator/Example.hs
@@ -12,8 +12,11 @@
 {-# LANGUAGE UndecidableInstances #-}
 
 module Example
-    ( Example.Toggle(..)
+    ( Example.Toggle_Aux(..)
+    , Example.Toggle(..)
+    , Example.Counter_Aux(..)
     , Example.Counter(..)
+    , Example.VarCounter_Aux(..)
     , Example.VarCounter(..)
     )
   where
@@ -22,6 +25,67 @@ import qualified HsBindgen.Runtime.Block as Block
 import qualified HsBindgen.Runtime.HasCField as HasCField
 import qualified HsBindgen.Runtime.Internal.Prelude as RIP
 
+{-| Auxiliary type used by 'Toggle'
+
+    __C declaration:__ @Toggle@
+
+    __defined at:__ @edge-cases\/iterator.h 3:16@
+
+    __exported by:__ @edge-cases\/iterator.h@
+-}
+newtype Toggle_Aux = Toggle_Aux
+  { unwrapToggle_Aux :: IO RIP.CBool
+  }
+  deriving stock (RIP.Generic)
+  deriving newtype (RIP.HasFFIType)
+
+-- __unique:__ @toToggle_Aux@
+foreign import ccall safe "wrapper" hs_bindgen_eca2bca8e63194be_base ::
+     IO RIP.Word8
+  -> IO (RIP.FunPtr (IO RIP.Word8))
+
+-- __unique:__ @toToggle_Aux@
+hs_bindgen_eca2bca8e63194be ::
+     Toggle_Aux
+  -> IO (RIP.FunPtr Toggle_Aux)
+hs_bindgen_eca2bca8e63194be =
+  \fun0 ->
+    fmap RIP.castFunPtrFromFFIType (hs_bindgen_eca2bca8e63194be_base (RIP.toFFIType fun0))
+
+-- __unique:__ @fromToggle_Aux@
+foreign import ccall safe "dynamic" hs_bindgen_703fc4bdc168721d_base ::
+     RIP.FunPtr (IO RIP.Word8)
+  -> IO RIP.Word8
+
+-- __unique:__ @fromToggle_Aux@
+hs_bindgen_703fc4bdc168721d ::
+     RIP.FunPtr Toggle_Aux
+  -> Toggle_Aux
+hs_bindgen_703fc4bdc168721d =
+  \funPtr0 ->
+    RIP.fromFFIType (hs_bindgen_703fc4bdc168721d_base (RIP.castFunPtrToFFIType funPtr0))
+
+instance RIP.ToFunPtr Toggle_Aux where
+
+  toFunPtr = hs_bindgen_eca2bca8e63194be
+
+instance RIP.FromFunPtr Toggle_Aux where
+
+  fromFunPtr = hs_bindgen_703fc4bdc168721d
+
+instance ( ty ~ IO RIP.CBool
+         ) => RIP.HasField "unwrapToggle_Aux" (RIP.Ptr Toggle_Aux) (RIP.Ptr ty) where
+
+  getField =
+    HasCField.fromPtr (RIP.Proxy @"unwrapToggle_Aux")
+
+instance HasCField.HasCField Toggle_Aux "unwrapToggle_Aux" where
+
+  type CFieldType Toggle_Aux "unwrapToggle_Aux" =
+    IO RIP.CBool
+
+  offset# = \_ -> \_ -> 0
+
 {-| __C declaration:__ @Toggle@
 
     __defined at:__ @edge-cases\/iterator.h 3:16@
@@ -29,12 +93,12 @@ import qualified HsBindgen.Runtime.Internal.Prelude as RIP
     __exported by:__ @edge-cases\/iterator.h@
 -}
 newtype Toggle = Toggle
-  { unwrapToggle :: Block.Block (IO RIP.CBool)
+  { unwrapToggle :: Block.Block Toggle_Aux
   }
   deriving stock (RIP.Generic)
   deriving newtype (RIP.HasFFIType)
 
-instance ( ty ~ Block.Block (IO RIP.CBool)
+instance ( ty ~ Block.Block Toggle_Aux
          ) => RIP.HasField "unwrapToggle" (RIP.Ptr Toggle) (RIP.Ptr ty) where
 
   getField =
@@ -43,7 +107,68 @@ instance ( ty ~ Block.Block (IO RIP.CBool)
 instance HasCField.HasCField Toggle "unwrapToggle" where
 
   type CFieldType Toggle "unwrapToggle" =
-    Block.Block (IO RIP.CBool)
+    Block.Block Toggle_Aux
+
+  offset# = \_ -> \_ -> 0
+
+{-| Auxiliary type used by 'Counter'
+
+    __C declaration:__ @Counter@
+
+    __defined at:__ @edge-cases\/iterator.h 10:14@
+
+    __exported by:__ @edge-cases\/iterator.h@
+-}
+newtype Counter_Aux = Counter_Aux
+  { unwrapCounter_Aux :: IO RIP.CInt
+  }
+  deriving stock (RIP.Generic)
+  deriving newtype (RIP.HasFFIType)
+
+-- __unique:__ @toCounter_Aux@
+foreign import ccall safe "wrapper" hs_bindgen_2202848aad97fe0a_base ::
+     IO RIP.Int32
+  -> IO (RIP.FunPtr (IO RIP.Int32))
+
+-- __unique:__ @toCounter_Aux@
+hs_bindgen_2202848aad97fe0a ::
+     Counter_Aux
+  -> IO (RIP.FunPtr Counter_Aux)
+hs_bindgen_2202848aad97fe0a =
+  \fun0 ->
+    fmap RIP.castFunPtrFromFFIType (hs_bindgen_2202848aad97fe0a_base (RIP.toFFIType fun0))
+
+-- __unique:__ @fromCounter_Aux@
+foreign import ccall safe "dynamic" hs_bindgen_73304cb84e9a2f8f_base ::
+     RIP.FunPtr (IO RIP.Int32)
+  -> IO RIP.Int32
+
+-- __unique:__ @fromCounter_Aux@
+hs_bindgen_73304cb84e9a2f8f ::
+     RIP.FunPtr Counter_Aux
+  -> Counter_Aux
+hs_bindgen_73304cb84e9a2f8f =
+  \funPtr0 ->
+    RIP.fromFFIType (hs_bindgen_73304cb84e9a2f8f_base (RIP.castFunPtrToFFIType funPtr0))
+
+instance RIP.ToFunPtr Counter_Aux where
+
+  toFunPtr = hs_bindgen_2202848aad97fe0a
+
+instance RIP.FromFunPtr Counter_Aux where
+
+  fromFunPtr = hs_bindgen_73304cb84e9a2f8f
+
+instance ( ty ~ IO RIP.CInt
+         ) => RIP.HasField "unwrapCounter_Aux" (RIP.Ptr Counter_Aux) (RIP.Ptr ty) where
+
+  getField =
+    HasCField.fromPtr (RIP.Proxy @"unwrapCounter_Aux")
+
+instance HasCField.HasCField Counter_Aux "unwrapCounter_Aux" where
+
+  type CFieldType Counter_Aux "unwrapCounter_Aux" =
+    IO RIP.CInt
 
   offset# = \_ -> \_ -> 0
 
@@ -54,12 +179,12 @@ instance HasCField.HasCField Toggle "unwrapToggle" where
     __exported by:__ @edge-cases\/iterator.h@
 -}
 newtype Counter = Counter
-  { unwrapCounter :: Block.Block (IO RIP.CInt)
+  { unwrapCounter :: Block.Block Counter_Aux
   }
   deriving stock (RIP.Generic)
   deriving newtype (RIP.HasFFIType)
 
-instance ( ty ~ Block.Block (IO RIP.CInt)
+instance ( ty ~ Block.Block Counter_Aux
          ) => RIP.HasField "unwrapCounter" (RIP.Ptr Counter) (RIP.Ptr ty) where
 
   getField =
@@ -68,7 +193,68 @@ instance ( ty ~ Block.Block (IO RIP.CInt)
 instance HasCField.HasCField Counter "unwrapCounter" where
 
   type CFieldType Counter "unwrapCounter" =
-    Block.Block (IO RIP.CInt)
+    Block.Block Counter_Aux
+
+  offset# = \_ -> \_ -> 0
+
+{-| Auxiliary type used by 'VarCounter'
+
+    __C declaration:__ @VarCounter@
+
+    __defined at:__ @edge-cases\/iterator.h 17:14@
+
+    __exported by:__ @edge-cases\/iterator.h@
+-}
+newtype VarCounter_Aux = VarCounter_Aux
+  { unwrapVarCounter_Aux :: RIP.CInt -> IO RIP.CInt
+  }
+  deriving stock (RIP.Generic)
+  deriving newtype (RIP.HasFFIType)
+
+-- __unique:__ @toVarCounter_Aux@
+foreign import ccall safe "wrapper" hs_bindgen_42a7337570f8b0d0_base ::
+     (RIP.Int32 -> IO RIP.Int32)
+  -> IO (RIP.FunPtr (RIP.Int32 -> IO RIP.Int32))
+
+-- __unique:__ @toVarCounter_Aux@
+hs_bindgen_42a7337570f8b0d0 ::
+     VarCounter_Aux
+  -> IO (RIP.FunPtr VarCounter_Aux)
+hs_bindgen_42a7337570f8b0d0 =
+  \fun0 ->
+    fmap RIP.castFunPtrFromFFIType (hs_bindgen_42a7337570f8b0d0_base (RIP.toFFIType fun0))
+
+-- __unique:__ @fromVarCounter_Aux@
+foreign import ccall safe "dynamic" hs_bindgen_43d902480175fccf_base ::
+     RIP.FunPtr (RIP.Int32 -> IO RIP.Int32)
+  -> RIP.Int32 -> IO RIP.Int32
+
+-- __unique:__ @fromVarCounter_Aux@
+hs_bindgen_43d902480175fccf ::
+     RIP.FunPtr VarCounter_Aux
+  -> VarCounter_Aux
+hs_bindgen_43d902480175fccf =
+  \funPtr0 ->
+    RIP.fromFFIType (hs_bindgen_43d902480175fccf_base (RIP.castFunPtrToFFIType funPtr0))
+
+instance RIP.ToFunPtr VarCounter_Aux where
+
+  toFunPtr = hs_bindgen_42a7337570f8b0d0
+
+instance RIP.FromFunPtr VarCounter_Aux where
+
+  fromFunPtr = hs_bindgen_43d902480175fccf
+
+instance ( ty ~ (RIP.CInt -> IO RIP.CInt)
+         ) => RIP.HasField "unwrapVarCounter_Aux" (RIP.Ptr VarCounter_Aux) (RIP.Ptr ty) where
+
+  getField =
+    HasCField.fromPtr (RIP.Proxy @"unwrapVarCounter_Aux")
+
+instance HasCField.HasCField VarCounter_Aux "unwrapVarCounter_Aux" where
+
+  type CFieldType VarCounter_Aux "unwrapVarCounter_Aux" =
+    RIP.CInt -> IO RIP.CInt
 
   offset# = \_ -> \_ -> 0
 
@@ -79,12 +265,12 @@ instance HasCField.HasCField Counter "unwrapCounter" where
     __exported by:__ @edge-cases\/iterator.h@
 -}
 newtype VarCounter = VarCounter
-  { unwrapVarCounter :: Block.Block (RIP.CInt -> IO RIP.CInt)
+  { unwrapVarCounter :: Block.Block VarCounter_Aux
   }
   deriving stock (RIP.Generic)
   deriving newtype (RIP.HasFFIType)
 
-instance ( ty ~ Block.Block (RIP.CInt -> IO RIP.CInt)
+instance ( ty ~ Block.Block VarCounter_Aux
          ) => RIP.HasField "unwrapVarCounter" (RIP.Ptr VarCounter) (RIP.Ptr ty) where
 
   getField =
@@ -93,6 +279,6 @@ instance ( ty ~ Block.Block (RIP.CInt -> IO RIP.CInt)
 instance HasCField.HasCField VarCounter "unwrapVarCounter" where
 
   type CFieldType VarCounter "unwrapVarCounter" =
-    Block.Block (RIP.CInt -> IO RIP.CInt)
+    Block.Block VarCounter_Aux
 
   offset# = \_ -> \_ -> 0